refactor(session): extract conflict-gate and dynamic-profile logic
Move conflict-gate evaluation (relevance scoring, cooldown decisions,
clarification resolution) into ConflictGate class in session-conflict-gate.ts.
Move dynamic-profile injection/stripping helpers into session-dynamic-profile.ts.

Session.ts: 2157 → 1990 lines (-167).
